+++ packages/powersync/lib/src/devtools/expose_credentials_connector.dart
@@ -9,7 +9,16 @@ final class ExposeCredentialsConnector extends PowerSyncBackendConnector {
   final PowerSyncDatabase database;
   final PowerSyncBackendConnector inner;
 
-  ExposeCredentialsConnector(this.inner, this.database);
+  factory ExposeCredentialsConnector(
+    PowerSyncBackendConnector inner,
+    PowerSyncDatabase database,
+  ) {
+    return inner is CustomCheckpointRequestConnector
+        ? _ExposeCustomWriteCheckpointsConnector(inner, database)
+        : ExposeCredentialsConnector._(inner, database);
+  }
+
+  ExposeCredentialsConnector._(this.inner, this.database);
 
   @override
   Future<PowerSyncCredentials?> fetchCredentials() async {
@@ -25,5 +34,21 @@ final class ExposeCredentialsConnector extends PowerSyncBackendConnector {
   @override
   Future<void> uploadData(PowerSyncDatabase database) {
     return inner.uploadData(database);
+  }
+}
+
+final class _ExposeCustomWriteCheckpointsConnector
+    extends ExposeCredentialsConnector
+    with CustomCheckpointRequestConnector {
+  final CustomCheckpointRequestConnector _custom;
+
+  _ExposeCustomWriteCheckpointsConnector(
+    this._custom,
+    PowerSyncDatabase database,
+  ) : super._(_custom, database);
+
+  @override
+  Future<String> postCheckpointRequest(String clientId, String requestId) {
+    return _custom.postCheckpointRequest(clientId, requestId);
   }
 }